% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_metadata.R
\name{add_to_metadata}
\alias{add_to_metadata}
\title{Add new items to metadata}
\usage{
add_to_metadata(metadata = meta, code, name, label = NA, value = NA,
  class = "numeric")
}
\arguments{
\item{metadata}{metadata data.frame}

\item{code}{variable code of the new var in data}

\item{name}{a more expressive variable name}

\item{label}{vector of labels for each value}

\item{value}{each value as a vector}

\item{class}{class of the new vector as \code{numeric}, \code{factor} or \code{character}}
}
\value{
data.frame
}
\description{
Add new items to metadata, effectively a dplyr::bind_rows on top of current metadata
}
\examples{
 \dontrun{
 add_to_metadata(metadata=meta, code = "sex", name = "Repondents sex", label=c("male","female"), value=c(1:2), class="factor")
 }

}
\author{
Markus Kainu <markuskainu@gmail.com>
}
