% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_labelled_data.R
\name{create_labelled_data}
\alias{create_labelled_data}
\title{Create a labeled data.frame for SPSS, Stata or SAS export}
\usage{
create_labelled_data(data, metadata)
}
\arguments{
\item{data}{a data.frame stripped from any attributes}

\item{metadata}{a data.frame with matching meta in tidy format ie. each value of each variable on its own row}
}
\value{
data.frame
}
\description{
Create a labeled data.frame from plain data.frame and matching metadata
}
\examples{
 \dontrun{
 create_labelled_data(data = stripped_spss_data, metadata = meta_spss)
 }

}
\author{
Markus Kainu <markuskainu@gmail.com>
}
