% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{create_metadata}
\alias{create_metadata}
\title{Create a metadata from labeled data.frame SPSS, Stata or SAS export}
\usage{
create_metadata(data)
}
\arguments{
\item{data}{a labeled data.frame imported from Stata, SPSS or SAS using haven}
}
\value{
data.frame
}
\description{
Create a metadata from labeled data.frame imported from SPSS, Stata or SAS file
}
\examples{
 \dontrun{
 create_metadata(data = imported_spss_data)
 }

}
\author{
Markus Kainu <markuskainu@gmail.com>
}
