% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_data.R
\name{label_data}
\alias{label_data}
\title{Create a string or factor vector based on value labels in matching metadata}
\usage{
label_data(data = d, variable.data = varcode,
  variable.meta = variable.data, metadata = meta, into.factor = TRUE)
}
\arguments{
\item{data}{A data.frame including the source variable to be labeled}

\item{variable.data}{Code of the source variable in data.frame}

\item{variable.meta}{Code of the source variable in the matching metadata}

\item{metadata}{A tidy metadata}

\item{into.factor}{Whether the output variable should be converted into a factor or not}
}
\value{
vector
}
\description{
Create a string or factor vector based on value labels in matching metadata
}
\examples{
 \dontrun{
 label_data(data=d, variable.data="sex", variable.meta=variable.data, metadata=meta, into.factor=TRUE)
 }

}
\author{
Markus Kainu <markuskainu@gmail.com>
}
