% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_attributes.R
\name{strip_attributes}
\alias{strip_attributes}
\title{Strip attributes from labelled data.frame using as.vector() function}
\usage{
strip_attributes(data)
}
\arguments{
\item{data}{Labelled data.frame}
}
\value{
data.frame
}
\description{
Strip attributes from labelled data.frame using as.vector() function
}
\examples{
 \dontrun{
 strip_attributes(data = spss_data)
 }

}
\author{
Markus Kainu <markuskainu@gmail.com>
}
